Addressed PR comments - Part 1 (errors)
