--- tools/cli/internal/openapi/errors/param_conflict_error.go
@@ -1,25 +1,0 @@
-// Copyright 2024 MongoDB Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package errors
-
-import "fmt"
-
-type ParamConflictError struct {
-	Entry string
-}
-
-func (e ParamConflictError) Error() string {
-	return fmt.Sprintf("there was a conflict with the Param component: %s", e.Entry)
-}
